:truck: chore(api) move knowledge repositories to `src`
