+++ api/lib/infrastructure/repositories/certifiable-profile-for-learning-content-repository.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 
 import { knex } from '../../../db/knex-database-connection.js';
 import { CertifiableProfileForLearningContent } from '../../../src/shared/domain/models/CertifiableProfileForLearningContent.js';
-import * as knowledgeElementRepository from './knowledge-element-repository.js';
+import * as knowledgeElementRepository from '../../../src/shared/infrastructure/repositories/knowledge-element-repository.js';
 
 const get = async function ({ id, profileDate, learningContent }) {
   const knowledgeElements = await knowledgeElementRepository.findUniqByUserId({ userId: id, limitDate: profileDate });
