+++ api/src/prescription/campaign-participation/infrastructure/repositories/campaign-participation-result-repository.js
@@ -1,9 +1,9 @@
 import * as campaignRepository from '../../../../../lib/infrastructure/repositories/campaign-repository.js';
-import * as knowledgeElementRepository from '../../../../../lib/infrastructure/repositories/knowledge-element-repository.js';
 import { CampaignParticipationResult } from '../../../../shared/domain/models/CampaignParticipationResult.js';
 import * as areaRepository from '../../../../shared/infrastructure/repositories/area-repository.js';
 import * as assessmentRepository from '../../../../shared/infrastructure/repositories/assessment-repository.js';
 import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
+import * as knowledgeElementRepository from '../../../../shared/infrastructure/repositories/knowledge-element-repository.js';
 import * as campaignParticipationRepository from './campaign-participation-repository.js';
 
 const campaignParticipationResultRepository = {
